requirements.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.12.4
flake8==5.0.4
importlib-metadata==4.2.0
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.4.0
zipp==3.13.0
